+++ bert_project/bert_project/settings.py
@@ -56,7 +56,7 @@ ROOT_URLCONF = 'bert_project.urls'
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [],
+        'DIRS': [BASE_DIR/ 'templates', ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
