+++ bert_project/main/templates/main/detail.html
@@ -1,0 +1,41 @@
+{% extends 'base.html' %}
+{% block content %}
+      <div class="container">
+        <div class="row my-3">
+          <div class="col-md-6 col-lg-5">
+            <h1>리뷰</h1>
+      <!-- Blog post-->
+        <a href="#!"><img class="card-img-top" src="./images/smail.png" alt="..." /></a>
+        <div class="card-body">
+            <h2 class="card-title h4">냉장고</h2>
+            <p class="card-text">리뷰가 좋아요</p>
+        </div>
+    </div>
+
+
+
+    <div class="container">
+      <div class="row my-3">
+        <div class="col-md-6 col-lg-5">
+          <h1>리뷰</h1>
+    <!-- Blog post-->
+      <a href="#!"><img class="card-img-top" src="./images/smail.png" alt="..." /></a>
+      <div class="card-body">
+          <h2 class="card-title h4">냉장고</h2>
+          <p class="card-text">리뷰가 좋아요</p>
+      </div>
+    </div>
+
+    <div class="container">
+      <div class="row my-3">
+        <div class="col-md-6 col-lg-5">
+          <h1>리뷰</h1>
+    <!-- Blog post-->
+      <a href="#!"><img class="card-img-top" src="./images/smail.png" alt="..." /></a>
+      <div class="card-body">
+          <h2 class="card-title h4">냉장고</h2>
+          <p class="card-text">리뷰가 좋아요</p>
+      </div>
+  </div>
+
+  {% endblock content %}
