+++ bert_project/main/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from . import views
+
+app_name = 'main'
+
+urlpatterns = [
+    path('', views.index, name='index'),
+    path('detail/', views.detail, name='detail'),
+]
