--- project_django/bert_project/main/templates/main/base.html
@@ -1,19 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html lang='ko'>
-
-  <head>
-    <meta charset='UTF-8'>
-    <title>Sentiment-Analysis</title>
-    <link rel='stylesheet' type='text/css' href="{% static 'main/custom_style.css' %}" />
-  </head>
-
-  <body>
-
-    {% block content_1 %}
-    {% endblock %}
-
-  </body>
-
-</html>
