--- project_django/requirements.txt
@@ -1,61 +1,0 @@
-absl-py==0.15.0
-asgiref==3.6.0
-astunparse==1.6.3
-cachetools==5.3.0
-certifi==2022.12.7
-charset-normalizer==3.1.0
-click==8.1.3
-colorama==0.4.6
-Django==3.2
-filelock==3.12.0
-flatbuffers==1.12
-gast==0.4.0
-google-auth==2.17.3
-google-auth-oauthlib==1.0.0
-google-pasta==0.2.0
-grpcio==1.34.1
-h5py==3.1.0
-huggingface-hub==0.0.12
-idna==3.4
-importlib-metadata==6.5.0
-joblib==1.2.0
-keras-nightly==2.5.0.dev2021032900
-Keras-Preprocessing==1.1.2
-Markdown==3.4.3
-MarkupSafe==2.1.2
-numpy==1.19.5
-oauthlib==3.2.2
-opt-einsum==3.3.0
-packaging==23.1
-pandas==1.2.4
-protobuf==4.22.3
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-python-dateutil==2.8.2
-pytz==2023.3
-PyYAML==6.0
-regex==2023.3.23
-requests==2.28.2
-requests-oauthlib==1.3.1
-rsa==4.9
-sacremoses==0.0.53
-sentencepiece==0.1.96
-six==1.15.0
-sqlparse==0.4.4
-tensorboard==2.12.2
-tensorboard-data-server==0.7.0
-tensorboard-plugin-wit==1.8.1
-tensorflow==2.5.0
-tensorflow-addons==0.17.0
-tensorflow-estimator==2.5.0
-termcolor==1.1.0
-tokenizers==0.10.3
-torch==1.11.0
-tqdm==4.65.0
-transformers==4.8.2
-typeguard==3.0.2
-typing-extensions==4.5.0
-urllib3==1.26.15
-Werkzeug==2.2.3
-wrapt==1.12.1
-zipp==3.15.0
