+++ requirements.txt
@@ -1,10 +1,15 @@
-asgiref==3.6.0
-backports.zoneinfo==0.2.1
-Django==4.2
-numpy==1.24.2
-pandas==2.0.0
-python-dateutil==2.8.2
-pytz==2023.3
-six==1.16.0
-sqlparse==0.4.3
-tzdata==2023.3
+numpy==1.19.5
+requests==2.28.2
+tqdm==4.65.0
+django==3.2 
+tensorflow==2.5.0
+tensorflow-addons==0.17.0
+transformers==4.8.2
+sentencepiece==0.1.96
+torch==1.11.0
+pandas==1.2.4
+
+
+
+
+
